refactor tests for faster evaluation
