+++ Tests/Unit/test_complex.py
@@ -38,77 +38,52 @@ def test_complex_with_polymer():
     truth_o = OrderedPolymerSpecies([a, Complex([b,a], ordered = True),a])
     assert d_co.parent==truth_o
 
-    #Cannot complex two Species inside a PolymerSpecies without puting the PolymerSpecies into a Conformation first
-    with pytest.raises(TypeError):
-        c = Complex([Species("S"), d[0], d[2]])
-
-
-def test_complex_with_single_polymer():
+def test_complex_a_whole_polymer():
     a = Species('A')
     b = Species('B')
     p = OrderedPolymerSpecies([a,b,a])
+    c = Complex([Species("S"), p[0], p[2]])
+    pc = c.parent
 
-    #This should just produce a ComplexSpecies around the PolymerSpecies
+    #This works as expected
     c2 = Complex([p, Species("S")])
     assert c2 == ComplexSpecies([p, Species("S")], called_from_complex = True)
 
     #If the Polymer is in a Conformation, it cannot be Complexed.
-    pc = PolymerConformation(polymer = p)
-    assert str(pc.polymers[0]) == str(p)
-    assert pc.polymers[0].parent == pc
-    assert p.parent is None
-
-    with pytest.raises(ValueError):
+    with pytest.raises(TypeError):
         c2 = Complex([pc.polymers[0], Species("S")])
-
-    #A monomer form the polymer can still be complexed, however
-    c2 = Complex([pc.polymers[0][0], Species("S")])
-    assert isinstance(c2.parent, PolymerConformation)
-    assert c2.parent != pc
-    assert len(c2.parent.complexes) == 1
-    assert len(c2.parent.polymers) == 1
-    assert str(c2) == str(ComplexSpecies([pc.polymers[0][0], Species("S")], called_from_complex = True))
-
+    
 
 def test_complex_with_multiple_polymers():
     a = Species('A')
     b = Species('B')
     p = OrderedPolymerSpecies([a,b,a])
-    pc0 = PolymerConformation(polymer = p)
     p2 = OrderedPolymerSpecies([b, a, b])
-    pc2 = PolymerConformation(polymer = p2)
 
-    #Bind one two monomers from one polymer
-    #Polymers must be placed into a Conformation before being bound together
-    with pytest.raises(TypeError):
-        c = Complex([Species("S"), p[0], p[2]])
-
-    c = Complex([Species("S"), pc0.polymers[0][0], pc0.polymers[0][2]])
-
+    #Bind a single OrderedPolymerSpecies into a PolymerConformation
+    c = Complex([Species("S"), p[0], p[2]])
     #Correct parent
     assert c.parent == PolymerConformation([ComplexSpecies([Species("S"), p[0], p[2]], called_from_complex = True)])
     #correct complex returned
     assert str(c) == str(ComplexSpecies([Species("S"), p[0], p[2]], called_from_complex = True)) 
 
     #Ordered Case
-    oc = Complex([Species("S"), pc0.polymers[0][0], pc0.polymers[0][2]], ordered = True)
+    oc = Complex([Species("S"), p[0], p[2]], ordered = True)
     #Correct parent
     assert oc.parent == PolymerConformation([OrderedComplexSpecies([Species("S"), p[0], p[2]], called_from_complex = True)])
     #correct complex returned
     assert str(oc) == str(OrderedComplexSpecies([Species("S"), p[0], p[2]], called_from_complex = True))
 
     #Two polymers which bind together
-    #Polymers must be placed into a Conformation before being bound together
-    with pytest.raises(TypeError):
-        c2 = Complex([Species("S"), p[0], p2[1]])
-
-    c2 = Complex([Species("S"), pc0.polymers[0][0], pc2.polymers[0][1]]) 
+    c2 = Complex([Species("S"), p[0], p2[1]])
     assert c2.parent == PolymerConformation([ComplexSpecies([Species("S"), p[0], p2[1]], called_from_complex = True)])
     assert str(c2) == str(ComplexSpecies([Species("S"), p[0], p2[1]], called_from_complex = True))
 
     #Two polymers already bound together in a Conformation interacting at a new location
     l1 = c2.parent.polymers[0][1]
+    print(l1, l1.parent, l1.parent.parent)
     l2 = c2.parent.polymers[1][0]
+    print(l2, l2.parent, l2.parent.parent)
     c3 = Complex([l1, l2], ordered = True)
     assert c3.parent == PolymerConformation([
         ComplexSpecies([Species("S"), p[0], p2[1]], called_from_complex = True),
@@ -117,31 +92,20 @@ def test_complex_with_multiple_polymers():
     assert str(c3) == str(OrderedComplexSpecies([p[1], p2[0]], called_from_complex = True))
 
 def test_complex_with_polymer_replacement():
-    #These tests show how the order of binding can matter. 
-    #Best practices is to put everything into a PolymerConformation before doing Complex if PolymerConformations are being used.
-
+    #This occurs when a Complex is formed around a Monomer in a Polymer in a PolymerConformation which does not include any other Monomers with parents.
     a = Species('A')
     b = Species('B')
-    s = Species("S")
     p = OrderedPolymerSpecies([a,b,a])
-    pc0 = PolymerConformation(polymer = p) #Put the polymer inside a conformation
-    #Bind two monomers together
-    pc = Complex([pc0.polymers[0][0], pc0.polymers[0][1]]).parent #get a PolymerConformation
-
-    assert isinstance(pc, PolymerConformation)
-
+    pc = Complex([p[0], p[1]]).parent #get a PolymerConformation
     #create a Complex around an unbound element of p (from within pc)
-    c = Complex([s, pc.polymers[0][2], s], ordered = True)
-    assert str(c) == str(OrderedComplexSpecies([s, p[2], s], called_from_complex = True))
-    assert str(pc.polymers[0]) == str(c.parent.polymers[0])
-    assert len(c.parent.complexes) > len(pc.complexes)
+    c = Complex([Species("S"), pc.polymers[0][2], Species("S")], ordered = True)
+    assert str(c) == str(OrderedComplexSpecies([Species("S"), p[2], Species("S")], called_from_complex = True))
 
     #Make the same thing with the replacement first
-    #this gives a different final complex than doing things in the previous order
-    p2 = OrderedPolymerSpecies([a,b,Complex([s, a, s], ordered = True)])
-    pc2 = PolymerConformation(polymer = p2)
-    assert str(c.parent) != str(p2)
-    assert c.parent.parent != Complex([pc2.polymers[0][0], pc2.polymers[0][1]]).parent
+    p2 = OrderedPolymerSpecies([a,b,Complex([Species("S"), a, Species("S")], ordered = True)])
+    assert str(c.parent) == str(p2)
+    assert c.parent.parent == Complex([p2[0], p2[1]]).parent
+    
 
 def test_complex_with_a_complex_in_a_conformation():
     #This occurs when Complexes are formed around Complexes in PolymerConformations.
@@ -151,14 +115,13 @@ def test_complex_with_a_complex_in_a_conformation():
     b = Species('B')
     c = Species('C')
     p = OrderedPolymerSpecies([a,b,c])
-    pc0 = PolymerConformation(polymer = p)
-    pc = Complex([pc0.polymers[0][0], pc0.polymers[0][1]], ordered = True).parent #get a PolymerConformation
+    pc = Complex([p[0], p[1]], ordered = True).parent #get a PolymerConformation
     c = pc.complexes[0] #get a complex from the PolymerConformation
 
     c2 = Complex([c, Species("S")], ordered = True) #Create a Complex with a Complex
     pc2 = c2.parent
     assert str(c2) == str(OrderedComplexSpecies([p[0], p[1], Species("S")])) #merging done correctly
-    assert pc2 == Complex([pc0.polymers[0][0], pc0.polymers[0][1], Species("S")], ordered = True).parent #check the parent PolymerConformation
+    assert pc2 == Complex([p[0], p[1], Species("S")], ordered = True).parent #check the parent PolymerConformation
     assert len(pc2.complexes) == 1
 
     #Create a PolymerConformation with two complexes
@@ -170,8 +133,8 @@ def test_complex_with_a_complex_in_a_conformation():
     #merge the two complexes in pc3
     c4 = Complex([pc3.complexes[0], pc3.complexes[1], Species("S3")], ordered = True)
     assert len(c4.parent.complexes) == 1
-    assert str(c4) == str(Complex([pc0.polymers[0][0], pc0.polymers[0][1], Species("S"), pc0.polymers[0][0], pc0.polymers[0][2], Species("S2"), Species("S3")], ordered = True))
-    assert c4.parent == Complex([pc0.polymers[0][0], pc0.polymers[0][1], Species("S"), pc0.polymers[0][0], pc0.polymers[0][2], Species("S2"), Species("S3")], ordered = True).parent
+    assert str(c4) == str(Complex([p[0], p[1], Species("S"), p[0], p[2], Species("S2"), Species("S3")], ordered = True))
+    assert c4.parent == Complex([p[0], p[1], Species("S"), p[0], p[2], Species("S2"), Species("S3")], ordered = True).parent
 
     
 def test_invalid_complex():
