+++ Tests/Unit/test_plotting.py
@@ -1,16 +1,13 @@
+
 #  Copyright (c) 2020, Build-A-Cell. All rights reserved.
 #  See LICENSE file in the project root directory for details.
 
-import copy
-import re
-import warnings
-
-import pytest
-
-import biocrnpyler as bcp
-from biocrnpyler import CRNPlotter, DNA_construct, Promoter,IntegraseSite,\
-    RBS,CDS,Terminator, Complex, Species, DNA_part, RegulatedPromoter
 from biocrnpyler.dna_construct import RNA_construct
+import pytest
+#from unittest import TestCase
+from biocrnpyler import CRNPlotter, DNA_construct, Promoter,IntegraseSite,\
+                        RBS,CDS,Terminator, Complex, Species, DNA_part, RegulatedPromoter
+import copy
 
 def test_CRNPlotter():
     class dummy_renderer:
@@ -28,7 +25,7 @@ def test_CRNPlotter():
             ax.plot([0],[1])
 
             return(0,100)
-
+    
     d_r = dummy_renderer()
     test_plotter = CRNPlotter(dna_renderer=dummy_renderer(),rna_renderer=dummy_renderer(),cmap=[0,1,2,3,4,5,6])
 
@@ -80,7 +77,7 @@ def test_CRNPlotter():
     boundspec = test_plotter.make_dpl_from_species(Species("ooga",material_type="booga")) #make a species to be bound to the promoter
     revpart = test_plotter.part_dpl_dict[RegulatedPromoter("p1",regulators=["r1","r2"])].get_directed("reverse",bound=[boundspec]) #return reverse promoter with things bound
     assert(revpart.get_dpl()[0]["type"]=="Operator") #reverse promoter has operators on the left
-
+    
 
     test_plotter.colordict = {"Promoter":"purple"}
     prom_dpl = test_plotter.make_dpl_from_part(Promoter("pconst")) #make a promoter part
@@ -97,7 +94,7 @@ def test_CRNPlotter():
     test_construct4 = DNA_construct([RegulatedPromoter("p1",regulators=["r1","r2"]),DNA_part("test")])
     test_plotter.make_dpl_from_species(test_construct4.get_species())
     assert(test_construct4.get_species() in test_plotter.species_dpl_dict) #construct is in the dictionary
-
+    
     rna_construct = RNA_construct([RBS("ooga"),CDS("booga")])
     condpl1 = test_plotter.make_dpls_from_construct(rna_construct)
     assert(condpl1.material_type == "rna") #material type is set correctly
@@ -106,7 +103,7 @@ def test_CRNPlotter():
     condpl2 = test_plotter.make_dpls_from_construct(rna_construct2)
     assert(condpl2==condpl1) #this construct is already in the dictionary! Just return it
 
-
+    
 
     bound_construct = Complex([test_construct4.get_species()[1],Species("integrase",material_type="protein")]).parent
     #making a construct that is bound to something
@@ -121,33 +118,5 @@ def test_CRNPlotter():
     #make sure the part inside the construct made it into the species dict
     assert(Species("mydna",material_type="dna") in test_plotter.species_dpl_dict)
 
+    
 
-def test_render_network_bokeh():
-    # Create a simple reaction
-    parameters = {
-        (None, None, "ktx"): 0.05,
-        (None, None, "kb"): 100,
-        (None, None, "ku"): 10,
-        (None, None, "ktl"): 0.05,
-        ("rna_degredation_mm", None, "kdeg"): 0.001,
-        (None, None, "cooperativity"): 2}
-    txtl = bcp.TxTlExtract('mixture1', parameters=parameters)
-    dna = bcp.DNAassembly(
-        'mydna', promoter=bcp.RegulatedPromoter('plac', ['laci']),
-        rbs='UTR1', protein='GFP', initial_concentration=10)
-    txtl.add_component(dna)
-    crn1 = txtl.compile_crn()
-    crn1.add_reactions([
-        bcp.Reaction.from_massaction(
-            [bcp.Species('mydna', material_type='rna')], [], k_forward=0.1)])
-
-    # Generate a graph and make sure it is created
-    with warnings.catch_warnings(record=True) as records:
-        plot = bcp.render_network_bokeh(crn1)
-
-    for w in records:
-        if re.search("plotting disabled", w.message.args[0]):
-            pytest.fail("plotting library import error")
-        else:
-            warnings.showwarning(
-                w.message.args[0], w.category, w.filename, w.lineno)
