+++ pytest.ini
@@ -4,5 +4,6 @@ minversion = 5.4
 ## addopts = #--nbval-lax --cov=biocrnpyler --durations=10
 ## addopts = --cov=biocrnpyler --durations=10
 testpaths =
-    Tests
+    Tests/Unit
+    # Tests/Combinatorial
 #    examples
